Create calculators for effects and se (mreg logistic / yreg logistic)

Construct functions for the conditional effect estimates and their standard errors in the mreg logistic / yreg logistic setting. Internally, this function deconstruct model objects and feed parameter estiamtes to the internal worker functions calc_myreg_mreg_logistic_yreg_logistic_est and calc_myreg_mreg_logistic_yreg_logistic_se.

calc_myreg_mreg_logistic_yreg_logistic(
  mreg,
  mreg_fit,
  yreg,
  yreg_fit,
  avar,
  mvar,
  cvar,
  interaction
)

Arguments

mreg
A character vector of length 1. Mediator regression type: "linear" or "logistic".

mreg_fit
Model fit from fit_mreg

yreg
A character vector of length 1. Outcome regression type: "linear", "logistic", "loglinear", "poisson", "negbin", "survCox", "survAFT_exp", or "survAFT_weibull".

yreg_fit
Model fit from fit_yreg

avar
A character vector of length 1. Treatment variable name.

mvar
A character vector of length 1. Mediator variable name.

cvar
A character vector of length > 0. Covariate names. Use NULL if there is no covariate. However, this is a highly suspicious situation. Even if avar is randomized, mvar is not. Thus, there should usually be some confounder(s) to account for the common cause structure (confounding) between avar and yvar.

interaction
A logical vector of length 1. Default to TRUE. Whether to include a mediator-treatment interaction term in the outcome regression model.

Value

A list contraining a function for effect estimates and a function for corresponding standard errors.