Summary with robust sandwich variance estimator for modified Poisson

This is a version of summary.glm modified to use the robust variance estimator sandwich.

# S3 method for regmedint_mod_poisson
summary(object, ...)

Arguments

object

A model object of the class regmedint_mod_poisson

...

For compatibility with the generic.

Value

An object of the class summary.glm