Create a vector of coefficients from the outcome model (yreg)

This function extracts coef from yreg_fit and 3s with zeros appropriately to create a named vector consistently having the following elements: (Intercept) (a zero element is added for yreg = "survCox" for which no intercept is estimated (the baseline hazard is left unspecified)), avar, mvar, avar:mvar (a zero element is added when interaction = FALSE). cvar (this part is eliminated when cvar = NULL), EMM_AC_Ymodel (this part is eliminated when EMM_AC_Ymodel = NULL), EMM_MC (this part is eliminated when EMM_MC = NULL).

theta_hat(
  yreg,
  yreg_fit,
  avar,
  mvar,
  cvar,
  EMM_AC_Ymodel = NULL,
  EMM_MC = NULL,
  interaction
)

Arguments

yreg

A character vector of length 1. Outcome regression type: "linear", "logistic", "loglinear", "poisson", "negbin", "survCox", "survAFT_exp", or "survAFT_weibull".

yreg_fit

Model fit object for yreg (outcome model).

avar

A character vector of length 1. Treatment variable name.

mvar

A character vector of length 1. Mediator variable name.

cvar

A character vector of length > 0. Covariate names. Use NULL if there is no covariate. However, this is a highly suspicious situation. Even if avar is randomized, mvar is not. Thus, there are usually some confounder(s) to account for the common cause structure (confounding) between mvar and yvar.

EMM_AC_Ymodel

A character vector of length > 0. Effect modifiers names. The covariate vector in treatment-covariate product term in the outcome model.

EMM_MC

A character vector of length > 0. Effect modifiers names. The covariate vector in mediator-covariate product term in outcome model.

interaction

A logical vector of length 1. The presence of treatment-mediator interaction in the outcome model. Default to TRUE.

Value

A named numeric vector of coefficients.