regmedint: A package for regression-based causal mediation analysis

The package is a simple R implementation of the SAS macro as described in Valeri & VanderWeele 2013 and Valeri & VanderWeele 2015 https://www.hsph.harvard.edu/tyler-vanderweele/tools-and-tutorials/.

This is a user-interface for regression-based causal mediation analysis as described in Valeri & VanderWeele 2013 and Valeri & VanderWeele 2015.

regmedint(
  data,
  yvar,
  avar,
  mvar,
  cvar,
  a0,
  a1,
  m_cde,
  c_cond,
  mreg,
  yreg,
  interaction = TRUE,
  casecontrol = FALSE,
  eventvar = NULL
)

Arguments

data
Data frame containing the relevant variables.

yvar
A character vector of length 1. Outcome variable name. It should be the time variable for survival outcomes.

avar
A character vector of length 1. Treatment variable name.

mvar
A character vector of length 1. Mediator variable name.

cvar
A character vector of length > 0. Covariate names. Use NULL if there is no covariate. However, this is a highly suspicious situation. Even if avar is randomized, mvar is not. Thus, there should usually be some confounder(s) to account for the common cause structure (confounding) between avar and yvar.

a0
A numeric vector of length 1. Reference level of treatment variable that is considered "untreated" or "unexposed".

a1
A numeric vector of length 1.

m_cde
A numeric vector of length 1. Mediator level at which controlled direct effect is evaluated at.

c_cond
A numeric vector of the same length as cvar. Required for the ful output. The conditional effect

mreg
A character vector of length 1. Outcome regression type. One of "linear"

yreg
A character vector of length 1. Outcome regression type. One of

interaction
A logical vector of length 1. Default to TRUE. Whether to include a mediator-treatment interaction term in the outcome regression model.

casecontrol
A logical vector of length 1. Default to FALSE. Whether data comes from a case-control study.

eventvar
An character vector of length 1. Only required for survival outcome regression models. Note that the coding is 1 for event and 0 for censoring, following the R survival package convention.

Value

regmedint object, which is a list containing the mediator regression object, the outcome regression object, and the regression-based mediation results.

Fitting models

Use the regmedint function to fit models and set up regression-based causal mediation analysis.

Examining results

Several methods are available to examine the regmedint object. print summary coef confint FIXME: Document once implemented.